Add rl framework

+++ backend/othello_agent/src/gameplay/constants.rs
@@ -1,8 +1,6 @@
 // create new variable for code characters
 
-use crate::gameplay::types::IPosition;
-
-use super::types::IBoard;
+use super::{ game::IBoard, position::IPosition };
 
 pub const CODE_CHARS: &str = "234567bcdfghjkmnpqrstvwxyz-";
 
@@ -86,9 +84,9 @@ pub const INITIAL_BOARD: IBoard = [
     [X, X, X, X, X, X, X, X],
 ];
 
-pub const DEFAULT_CORNER_SCORE: i16 = 12;
-pub const DEFAULT_EDGE_SCORE: i16 = 4;
-pub const DEFAULT_OTHER_SCORE: i16 = 1;
+pub const DEFAULT_CORNER_SCORE: i8 = 12;
+pub const DEFAULT_EDGE_SCORE: i8 = 4;
+pub const DEFAULT_OTHER_SCORE: i8 = 1;
 
 pub const DEFAULT_BOARD_WIDTH: usize = 8;
 pub const DEFAULT_BOARD_HEIGHT: usize = 8;
